* Merged in lots of stuff from my laptop:     - Preliminary sqlite3 integration for future use with editor.     - More bsptools work.  This still doesn't run, but it provides       pretty good documentation of how bsp files work.     - Support for drawing lines in 3D while running Quake 2. This       will be needed to debug AI, if we ever do AI.     - Other minor tweaks and fixes.     - A few preliminary warning fixes.  Much more to do.
git-svn-id: https://imlsrc.dartmouth.edu/svn/public/halyard/trunk@28192 b24aca02-f418-0410-9c49-835e4aa915c3

+++ Common/TPrimitives.cpp
@@ -161,7 +161,7 @@ TArgumentList &FIVEL_NS operator>>(TArgumentList &args, int16 &out)
     if (temp < MIN_INT16 || MAX_INT16 < temp)
 		throw TException(__FILE__, __LINE__,
 						 "Can't represent value as 16-bit integer");
-    out = temp;
+    out = static_cast<short>(temp);
 	args.LogTValueParameter(out);
     return args;
 }
